fix default theme typecast with type declaration file
