+++ libs/ui/src/types/theming.ts
@@ -1,45 +1,3 @@
-export type ButtonColorTargets = {
-  bg: string;
-  text: string;
-  border: string;
-  hoverBg: string;
-  hoverText: string;
-  hoverBorder: string;
-  focusBg: string;
-  focusText: string;
-  focusBorder: string;
-  activeBg: string;
-  activeText: string;
-  activeBorder: string;
-  disabledBg: string;
-  disabledText: string;
-  disabledBorder: string;
-};
+import { DefaultTheme } from 'styled-components';
 
-export type Theme = {
-  themeName: string;
-  bgColor: string;
-  fontColor: string;
-  primary: string;
-  error: string;
-  warning: string;
-  success: string;
-  info: string;
-  field: {
-    bg: string;
-    focus: string;
-    placeholderText: string;
-    disabled: string;
-  };
-  button: {
-    primary: ButtonColorTargets;
-    secondary: ButtonColorTargets;
-    tertiary: ButtonColorTargets;
-  };
-  tooltip: {
-    bg: string;
-  };
-  dropdown: {
-    bg: string;
-  };
-};
+export type Theme = DefaultTheme;
