+++ libs/ui/tsconfig.lib.json
@@ -2,7 +2,7 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "outDir": "../../dist/out-tsc",
-    "types": ["vite/client", "node", "@nrwl/react", "@nxext/react/client"]
+    "types": ["vite/client", "node", "@nrwl/react"]
   },
   "files": [
     "../../node_modules/@nrwl/react/typings/cssmodule.d.ts",
